+++ crates/sampo-core/src/release.rs
@@ -939,199 +939,3 @@ fn validate_fixed_dependencies(cfg: &Config, ws: &Workspace) -> Result<(), Strin
     }
     Ok(())
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn formats_single_dependency_update() {
-        let updates = vec![DependencyUpdate {
-            name: "pkg1".to_string(),
-            new_version: "1.2.0".to_string(),
-        }];
-        let msg = format_dependency_updates_message(&updates).unwrap();
-        assert_eq!(msg, "Updated dependencies: pkg1@1.2.0");
-    }
-
-    #[test]
-    fn formats_multiple_dependency_updates() {
-        let updates = vec![
-            DependencyUpdate {
-                name: "pkg1".to_string(),
-                new_version: "1.2.0".to_string(),
-            },
-            DependencyUpdate {
-                name: "pkg2".to_string(),
-                new_version: "2.0.0".to_string(),
-            },
-        ];
-        let msg = format_dependency_updates_message(&updates).unwrap();
-        assert_eq!(msg, "Updated dependencies: pkg1@1.2.0, pkg2@2.0.0");
-    }
-
-    #[test]
-    fn returns_none_for_empty_updates() {
-        let updates = vec![];
-        let msg = format_dependency_updates_message(&updates);
-        assert_eq!(msg, None);
-    }
-
-    #[test]
-    fn builds_dependency_updates_from_tuples() {
-        let tuples = vec![
-            ("pkg1".to_string(), "1.2.0".to_string()),
-            ("pkg2".to_string(), "2.0.0".to_string()),
-        ];
-        let updates = build_dependency_updates(&tuples);
-        assert_eq!(updates.len(), 2);
-        assert_eq!(updates[0].name, "pkg1");
-        assert_eq!(updates[0].new_version, "1.2.0");
-        assert_eq!(updates[1].name, "pkg2");
-        assert_eq!(updates[1].new_version, "2.0.0");
-    }
-
-    #[test]
-    fn creates_dependency_update_entry() {
-        let updates = vec![DependencyUpdate {
-            name: "pkg1".to_string(),
-            new_version: "1.2.0".to_string(),
-        }];
-        let (msg, bump) = create_dependency_update_entry(&updates).unwrap();
-        assert_eq!(msg, "Updated dependencies: pkg1@1.2.0");
-        assert_eq!(bump, Bump::Patch);
-    }
-
-    #[test]
-    fn creates_fixed_dependency_policy_entry() {
-        let (msg, bump) = create_fixed_dependency_policy_entry(Bump::Major);
-        assert_eq!(msg, "Bumped due to fixed dependency group policy");
-        assert_eq!(bump, Bump::Major);
-
-        let (msg, bump) = create_fixed_dependency_policy_entry(Bump::Minor);
-        assert_eq!(msg, "Bumped due to fixed dependency group policy");
-        assert_eq!(bump, Bump::Minor);
-    }
-
-    #[test]
-    fn infers_bump_from_version_changes() {
-        assert_eq!(infer_bump_from_versions("1.0.0", "2.0.0"), Bump::Major);
-        assert_eq!(infer_bump_from_versions("1.0.0", "1.1.0"), Bump::Minor);
-        assert_eq!(infer_bump_from_versions("1.0.0", "1.0.1"), Bump::Patch);
-
-        // Edge cases
-        assert_eq!(infer_bump_from_versions("0.1", "0.2"), Bump::Patch);
-        assert_eq!(infer_bump_from_versions("invalid", "1.0.0"), Bump::Patch);
-    }
-
-    #[test]
-    fn detect_all_dependency_explanations_comprehensive() {
-        use crate::types::{CrateInfo, Workspace};
-        use std::collections::BTreeSet;
-        use std::path::PathBuf;
-
-        // Create test workspace with dependencies
-        let ws = Workspace {
-            root: PathBuf::from("/test"),
-            members: vec![
-                CrateInfo {
-                    name: "pkg-a".to_string(),
-                    version: "1.0.0".to_string(),
-                    path: PathBuf::from("/test/pkg-a"),
-                    internal_deps: BTreeSet::from(["pkg-b".to_string()]),
-                },
-                CrateInfo {
-                    name: "pkg-b".to_string(),
-                    version: "1.0.0".to_string(),
-                    path: PathBuf::from("/test/pkg-b"),
-                    internal_deps: BTreeSet::new(),
-                },
-                CrateInfo {
-                    name: "pkg-c".to_string(),
-                    version: "1.0.0".to_string(),
-                    path: PathBuf::from("/test/pkg-c"),
-                    internal_deps: BTreeSet::new(),
-                },
-            ],
-        };
-
-        // Create config with fixed dependencies
-        let config = Config {
-            version: 1,
-            github_repository: None,
-            changelog_show_commit_hash: true,
-            changelog_show_acknowledgments: true,
-            fixed_dependencies: vec![vec!["pkg-a".to_string(), "pkg-c".to_string()]],
-            linked_dependencies: vec![],
-        };
-
-        // Create changeset that affects pkg-b only
-        let changesets = vec![ChangesetInfo {
-            packages: vec!["pkg-b".to_string()],
-            bump: Bump::Minor,
-            message: "feat: new feature".to_string(),
-            path: PathBuf::from("/test/.sampo/changesets/test.md"),
-        }];
-
-        // Simulate releases: pkg-a and pkg-c get fixed bump, pkg-b gets direct bump
-        let mut releases = BTreeMap::new();
-        releases.insert(
-            "pkg-a".to_string(),
-            ("1.0.0".to_string(), "1.1.0".to_string()),
-        );
-        releases.insert(
-            "pkg-b".to_string(),
-            ("1.0.0".to_string(), "1.1.0".to_string()),
-        );
-        releases.insert(
-            "pkg-c".to_string(),
-            ("1.0.0".to_string(), "1.1.0".to_string()),
-        );
-
-        let explanations = detect_all_dependency_explanations(&changesets, &ws, &config, &releases);
-
-        // pkg-a should have dependency update message (depends on pkg-b)
-        let pkg_a_messages = explanations.get("pkg-a").unwrap();
-        assert_eq!(pkg_a_messages.len(), 1);
-        assert!(pkg_a_messages[0]
-            .0
-            .contains("Updated dependencies: pkg-b@1.1.0"));
-        assert_eq!(pkg_a_messages[0].1, Bump::Patch);
-
-        // pkg-c should have fixed dependency policy message (no deps but in fixed group)
-        let pkg_c_messages = explanations.get("pkg-c").unwrap();
-        assert_eq!(pkg_c_messages.len(), 1);
-        assert_eq!(
-            pkg_c_messages[0].0,
-            "Bumped due to fixed dependency group policy"
-        );
-        assert_eq!(pkg_c_messages[0].1, Bump::Minor); // Inferred from version change
-
-        // pkg-b should have no messages (explicit changeset)
-        assert!(!explanations.contains_key("pkg-b"));
-    }
-
-    #[test]
-    fn detect_all_dependency_explanations_empty_cases() {
-        use crate::types::{CrateInfo, Workspace};
-        use std::collections::BTreeSet;
-        use std::path::PathBuf;
-
-        let ws = Workspace {
-            root: PathBuf::from("/test"),
-            members: vec![CrateInfo {
-                name: "pkg-a".to_string(),
-                version: "1.0.0".to_string(),
-                path: PathBuf::from("/test/pkg-a"),
-                internal_deps: BTreeSet::new(),
-            }],
-        };
-
-        let config = Config::default();
-        let changesets = vec![];
-        let releases = BTreeMap::new();
-
-        let explanations = detect_all_dependency_explanations(&changesets, &ws, &config, &releases);
-        assert!(explanations.is_empty());
-    }
-}
