+++ crates/sampo/src/release.rs
@@ -4,9 +4,5 @@ use std::io;
 
 pub fn run(args: &ReleaseArgs) -> io::Result<()> {
     let cwd = std::env::current_dir()?;
-    run_in(&cwd, args)
-}
-
-pub fn run_in(root: &std::path::Path, args: &ReleaseArgs) -> io::Result<()> {
-    run_release(root, args.dry_run)
+    run_release(&cwd, args.dry_run)
 }
